split test cases into unit and aggregation tests
